Implemented client logging methods

+++ client/Logging.h
@@ -10,7 +10,7 @@ void saveLogReport(LogReport message, bool verbose);
 void logFinalReport(int killCount);
 //void logParentInit();
 void logProcessMonitoringInit(char* processName, pid_t pid);
-void logProcessKill(pid_t pid, const char* name, unsigned long int duration);
-void logSelfDying(pid_t pid, const char* name, unsigned long int duration);
+void logProcessKill(pid_t pid, const char* name, uint32_t duration);
+void logSelfDying(pid_t pid, const char* name, uint32_t duration);
 
 #endif
